[Add] buildPrefabs / [Feat] UpgradeNodes / [Add] UpgradeUI_Ctrl
[Add] buildPrefabs / 건물 프리팹 추가
[Feat] UpgradeNodes / 재화 단위가 double, float, int로 나뉜거 int로 통합
[Add] UpgradeUI_Ctrl / 업그레이드 캔버스 on off

+++ Assets/Scripts/UpgradeNode/UpgradeTooltip.cs
@@ -40,8 +40,8 @@ public class UpgradeTooltip : MonoBehaviour
         {
             foreach (CostData costData in costs)
             {
-                // 공식: 기본비용 * (배율 ^ 현재레벨)
-                double currentPrice = costData.baseCost * Mathf.Pow(costData.costMultiplier, currentLevel);
+                // 기본비용 * (배율 ^ 현재레벨)
+                int currentPrice = (int)(costData.baseCost * Mathf.Pow(costData.costMultiplier, currentLevel));
 
                 // 텍스트 누적 (예: "Food 100\nWood 50\n")
                 tempString += $"\n{costData.currency} : {currentPrice:N0}";
